+++ HandheldCompanion/Devices/Lenovo/LegionGoSZ1.cs
@@ -29,8 +29,8 @@ namespace HandheldCompanion.Devices
             ];
             hidFilters = new()
             {
-                { 0xE310, new HidFilter(unchecked((short)0xFFA0), unchecked((short)0x0001)) }, // xinput
-                { 0xE311, new HidFilter(unchecked((short)0xFFA0), unchecked((short)0x0001)) }, // dinput
+                { 0xE310, new HidFilter(unchecked((short)0xFFA0), unchecked(0x0001)) }, // xinput
+                { 0xE311, new HidFilter(unchecked((short)0xFFA0), unchecked(0x0001)) }, // dinput
             };
 
             GyrometerAxis = new Vector3(-1.0f, 1.0f, 1.0f);
