+++ HandheldCompanion/Devices/Zotac/GamingZone.cs
@@ -349,7 +349,7 @@ namespace HandheldCompanion.Devices.Zotac
             data[11] = (byte)ModifierKeyID.LeftCtrl | (byte)ModifierKeyID.LeftWinCMD;
 
             // Keyboard: F11
-            data[13] = (byte)BtnKeyboardMapPos[KeyCode.F11];
+            data[13] = BtnKeyboardMapPos[KeyCode.F11];
 
             ushort CRC = CalcZotacCRC(data, 5, 62);
             data[63] = (byte)(CRC >> 8);
@@ -375,7 +375,7 @@ namespace HandheldCompanion.Devices.Zotac
             data[11] = (byte)ModifierKeyID.LeftCtrl | (byte)ModifierKeyID.LeftWinCMD;
 
             // Keyboard: F12
-            data[13] = (byte)BtnKeyboardMapPos[KeyCode.F12];
+            data[13] = BtnKeyboardMapPos[KeyCode.F12];
 
             ushort CRC = CalcZotacCRC(data, 5, 62);
             data[63] = (byte)(CRC >> 8);
@@ -553,10 +553,10 @@ namespace HandheldCompanion.Devices.Zotac
 
         private static ushort CalcFast(ushort seed, byte c)
         {
-            uint num1 = (uint)(((int)seed ^ (int)c) & 0xFF);
+            uint num1 = (uint)((seed ^ c) & 0xFF);
             uint num2 = num1 & 0x0F;
             int num3 = ((int)num2 << 4) ^ (int)num1;
-            uint num4 = (uint)((uint)num3 >> 4);
+            uint num4 = (uint)num3 >> 4;
             uint intermediate = (uint)(((num3 << 1) ^ (int)num4) << 4 ^ (int)num2);
             return (ushort)((((intermediate << 3) ^ num4) ^ ((uint)seed >> 8)) & 0xFFFF);
         }
@@ -699,8 +699,8 @@ namespace HandheldCompanion.Devices.Zotac
         {
             uint vRAM = (physicalInstalledRamGB + additionSize) * 4U;
 
-            openLibSys.WriteIoPortByte((ushort)112, (byte)122);
-            openLibSys.WriteIoPortByte((ushort)113, (byte)vRAM);
+            openLibSys.WriteIoPortByte(112, 122);
+            openLibSys.WriteIoPortByte(113, (byte)vRAM);
         }
 
         public override void SetFanControl(bool enable, int mode)
